Adding error handling

+++ Online-Store/src/main/java/com/store/Online/Store/service/impl/CommentServiceImpl.java
@@ -4,7 +4,9 @@ import com.store.Online.Store.dto.CommentRequest;
 import com.store.Online.Store.entity.Comment;
 import com.store.Online.Store.entity.Product;
 import com.store.Online.Store.entity.User;
+import com.store.Online.Store.exception.CommentNotFoundException;
 import com.store.Online.Store.exception.ProductNotFoundException;
+import com.store.Online.Store.exception.UserCreationException;
 import com.store.Online.Store.service.commentService;
 import com.store.Online.Store.repository.commentRepository;
 import com.store.Online.Store.service.productService;
@@ -12,8 +14,6 @@ import com.store.Online.Store.service.userService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -51,14 +51,14 @@ public class CommentServiceImpl implements commentService{
         if (userOptional.isPresent()) {
             comment.setUserId(userOptional.get());
         } else {
-             throw new IllegalArgumentException("user not found");
+             throw new UserCreationException("User not found with ID: " + userOptional);
         }
 
         Optional<Product> productOptional = productService.getProductById(commentRequest.getProductId());
         if (productOptional.isPresent()) {
             comment.setProductId(productOptional.get());
         } else {
-            throw new IllegalArgumentException("product not found");
+            throw new ProductNotFoundException("Product not found with ID: " + productOptional);
         }
 
         comment.setText(commentRequest.getText());
@@ -69,18 +69,30 @@ public class CommentServiceImpl implements commentService{
 
     @Override
     public void deleteImage(Long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new CommentNotFoundException("Comment not found with ID: " + commentId);
+        }
+
         commentRepository.deleteImageByCommentId(commentId);
     }
 
     @Override
-    public void deleteComment(long commentId) {
+    public void deleteComment(Long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw new CommentNotFoundException("Comment not found with ID: " + commentId);
+        }
+
         commentRepository.deleteById(commentId);
     }
 
     @Override
-   public List<Comment> getCommentsByProductIdSortedByRating(Long productId, Sort.Direction direction) {
-//        return commentRepository.findByProductIdOrderByRating(productId, direction);
-        return new ArrayList<>();
+    public List<Comment> getCommentsByProductIdSortedByRating(Long productId, Sort.Direction direction) {
+        Optional<Product> productOptional = productService.getProductById(productId);
+        Product product = productOptional.orElseThrow(() -> new ProductNotFoundException("Error id product"));
+
+        Sort sort = Sort.by(direction, "rating");
+
+        return commentRepository.findByProductId(product, sort);
     }
 
 }
